SDK: added specRunner tool
This tool uses the NPM module globby to find any files with a name that matches the glob '**/*.{spec,test}.js' and then either requires it to get global utils or adds it as a mocha test. After all matched files are either run or added, it executes mocha, providing globals for both cx-lang and the specRunner's own expect method.