Add base and self context keywords
The "this" and "super" context keywords are used to access a classe's instance methods and properties, but cannot access static properties or methods, and cannot be used within these.

I'm thinking of using "self" and "base" as static variants of "this" and "super" that are used to access static properties and methods of the current class, and should be useable from static or instance methods.